* fix, pre_tag_map: a memory leak was found when reloading rules containing   'filter' keywords. Thanks to Matt Jenkins for his support resolving the   issue. * fix, pre_tag_map: test introduced to check if pre_tag_map was defined   globally in the configuration. If yes, find_id_func() is only run once   per cycle for all defined plugins.

+++ src/pretag.h
@@ -190,16 +190,16 @@ struct id_table_index {
 struct id_table {
   char *filename;
   int type;
-  unsigned short int num;
+  unsigned int num;
   struct id_entry *ipv4_base;
-  unsigned short int ipv4_num;
+  unsigned int ipv4_num;
 #if defined ENABLE_IPV6
   struct id_entry *ipv6_base;
-  unsigned short int ipv6_num;
+  unsigned int ipv6_num;
 #endif
   struct id_entry *e;
   struct id_table_index index[MAX_ID_TABLE_INDEXES];
-  unsigned short int index_num;
+  unsigned int index_num;
   time_t timestamp;
   u_int32_t flags;
 };
